AppEngine-ify
