+++ .gitignore
@@ -1,3 +1,4 @@
 *.png
 .idea/
 .gitignore
+keys.go
