+++ LICENSE
@@ -1,4 +1,4 @@
-Copyright (c) 2014, Luke Granger-Brown
+Copyright (c) 2017, Luke Granger-Brown and Google Inc.
 All rights reserved.
 
 Redistribution and use in source and binary forms, with or without modification, are permitted provided that the following conditions are met:
