+++ templates/perform.html
@@ -202,7 +202,7 @@ var gogoGadget = function() {
 				},
 				dataType: 'json'
 			}).done(function(data) {
-				var skinUri = 'http://' + window.location.host + '/media/skin/' + data.filename + '.png';
+                var skinUri = data.url;
 				$('.change-skin a').attr('href', 'https://minecraft.net/profile/skin/remote?url=' + encodeURIComponent(skinUri));
 	
 				$('.change-skin').removeClass('hide');
